more code cleanup and reference update
