+++ skater/core/local_interpretation/__init__.py
@@ -1,5 +1,5 @@
 from .text_interpreter import *
 
-__all__ = [cleaner, relevance_wt_assigner, vectorize_as_tf_idf, get_feature_names, query_topk_with_feature_selection,
+__all__ = [relevance_wt_assigner, vectorize_as_tf_idf, get_feature_names, query_topk_with_feature_selection,
            query_topk_tfidf_features, query_topk_tfidf_features_in_doc, query_topk_tfidf_features_overall,
            convert_dataframe_to_dict, query_topk_tfidf_features_by_class, understand_estimator]
