+++ skater/util/__init__.py
@@ -1,9 +1,5 @@
 """Includes kernels, static type/properties for package, exceptions, logging."""
-# from .kernels import rbf_kernel
-# from .static_types import return_data_type, StaticTypes, ModelTypes, OutputTypes
-# from .data_structures import ControlledDict
-# from .model import get_predictor
-# from .serialization import pickle_method, unpickle_method
+
 
 __all__ = [
     'rbf_kernel',
